Add a section detailing CMW collection label usage
This is to address Issue#74. A CMW collection type exists to organize
individual attesters' results and not fracture the space of
representations of composite attester results. The topologies however
are still subject to Verifier scrutiny. If the topology is not present
in a measured boot chain signed by some other attester, and the only
trusted witness of the complete assembly of a system is the CMW
collection itself, we must be careful to ensure that the meaning of
labels is properly documented (with a profile) and integrity-protected
(with a signature).